% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowscols.R
\name{buildUnPivotControlTable}
\alias{buildUnPivotControlTable}
\title{Build a moveValuesToColumnsQ() control table that specifies a un-pivot.}
\usage{
buildUnPivotControlTable(nameForNewKeyColumn, nameForNewValueColumn,
  columnsToTakeFrom, ...)
}
\arguments{
\item{nameForNewKeyColumn}{character name of column to write new keys in.}

\item{nameForNewValueColumn}{character name of column to write new values in.}

\item{columnsToTakeFrom}{character array names of columns to take values from.}

\item{...}{not used, force later args to be by name}
}
\value{
control table
}
\description{
Some discussion and examples can be found here:
\url{https://winvector.github.io/replyr/articles/FluidData.html} and
here \url{https://github.com/WinVector/cdata}.
}
\examples{

buildUnPivotControlTable("measurmentType", "measurmentValue", c("c1", "c2"))

}
\seealso{
\code{\link[cdata]{moveValuesToRows}}, \code{\link{moveValuesToRowsQ}}
}
