% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinController.R
\name{convertDiagrameToPNG}
\alias{convertDiagrameToPNG}
\title{Convert a renderJoinDiagram \code{\link{renderJoinDiagram}} result to a PNG graphics item.}
\usage{
convertDiagrameToPNG(diagram, ..., pngFileName = NULL, tempDir = tempdir())
}
\arguments{
\item{diagram}{DiagrammeR::grViz result (please see replyr::renderJoinDiagram)}

\item{...}{force later arguments to bind by name.}

\item{pngFileName}{optional, file path where to save the PNG.}

\item{tempDir}{directory to create temporary files in (not deleted by this method).}
}
\value{
png rendering of diagram
}
\description{
Requires packages \code{DiagrammeR}, \code{htmlwidgets}, \code{webshot}, and \code{magick} properly installed to use.
Please see \code{vignette('DependencySorting', package = 'replyr')} and \code{vignette('joinController', package= 'replyr')} for more details.
}
\seealso{
\code{\link{renderJoinDiagram}}
}
