% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DebugFn.R
\name{DebugPrintFnE}
\alias{DebugPrintFnE}
\title{Capture arguments and environment of exception throwing function call for later debugging.}
\usage{
DebugPrintFnE(saveFile, fn, ...)
}
\arguments{
\item{saveFile}{path to save RDS to.}

\item{fn}{function to call}

\item{...}{arguments for fn}
}
\value{
fn(...) normally, but if fn(...) throws an exception save to saveFile RDS of list r such that do.call(r$fn,r$args) repeats the call to fn with args.
}
\description{
Run fn and print result, save arguments on failure.  Use on systems like ggplot()
where some calculation is delayed until print().
}
\examples{

d <- data.frame(x=1:5)
saveName <- paste0(tempfile('debug'),'.RDS')
f <- function(i) { (1:10)[[i]] }
# correct run
DebugPrintFnE(saveName, f, 5)
# now re-run
# capture error on incorrect run
tryCatch(
   DebugPrintFnE(saveName, f, 12),
   error = function(e) { print(e) })
# examine details
situation <- readRDS(saveName)
str(situation)
# fix and re-run
situation$args[[1]] <- 6
do.call(situation$fn, situation$args, envir=situation$env)
# clean up
file.remove(saveName)

}
\seealso{
\code{\link{DebugFn}}, \code{\link{DebugPrintFn}}, \code{\link{DebugFnE}}
}

