% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceName.R
\name{dplyr_src_to_db_handle}
\alias{dplyr_src_to_db_handle}
\title{Obsolete with dplyr 0.7.0 and forward}
\usage{
dplyr_src_to_db_handle(dplyr_src)
}
\arguments{
\item{dplyr_src}{remote data handle}
}
\value{
dplyr_src
}
\description{
Obsolete with dplyr 0.7.0 and forward
}
