% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinController.R
\name{example_employeeAndDate}
\alias{example_employeeAndDate}
\title{build some example tables}
\usage{
example_employeeAndDate(con)
}
\arguments{
\item{con}{db connection}
}
\value{
example tables
}
\description{
build some example tables
}
