% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinController.R
\name{executeLeftJoinPlan}
\alias{executeLeftJoinPlan}
\title{Execute a sequence of left joins.}
\usage{
executeLeftJoinPlan(tDesc, columnJoinPlan, ..., eagerCompute = TRUE,
  tempNameGenerator = makeTempNameGenerator("executeLeftJoinPlan"))
}
\arguments{
\item{tDesc}{description of tables, from \code{\link{tableDesription}} (and likely altered by user).}

\item{columnJoinPlan}{columns to join, from \code{\link{buildJoinPlan}} (and likely altered by user).}

\item{...}{force later arguments to bind by name.}

\item{eagerCompute}{logical if TRUE compute eager.}

\item{tempNameGenerator}{temp name generator produced by replyr::makeTempNameGenerator, used to record dplyr::compute() effects.}
}
\value{
joined table
}
\description{
Execute a sequence of left joins.
}
\examples{

d <- data.frame(id=1:3, weight= c(200, 140, 98))
tDesc <- rbind(tableDesription('d1', d),
               tableDesription('d2', d))
tDesc$keys[[1]] <- list(PrimaryKey= 'id')
tDesc$keys[[2]] <- list(PrimaryKey= 'id')
columnJoinPlan <- buildJoinPlan(tDesc)
executeLeftJoinPlan(tDesc, columnJoinPlan)

}
