% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.R
\name{\%land\%}
\alias{\%land\%}
\alias{\%land_\%}
\title{Assign or land a value to variable from a pipeline.}
\usage{
value \%land\% name

value \%land_\% name
}
\arguments{
\item{value}{value to write}

\item{name}{variable to write to}
}
\description{
land copies a pipeline value to a variable on the RHS, land_ copies a pipeline value to
a variable named by its RHS argument.  There is nothing these operators do
better than "->" and they are mostly just a proof of concept.
}
\examples{

library("dplyr")
7 \%>\% sin() \%land\% z1
7 \%>\% sin() \%land_\% 'z2'

}

