% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinController.R
\name{inspectDescrAndJoinPlan}
\alias{inspectDescrAndJoinPlan}
\title{check that a join plan is consistent with table descriptions}
\usage{
inspectDescrAndJoinPlan(tDesc, columnJoinPlan)
}
\arguments{
\item{tDesc}{description of tables, from \code{\link{tableDesription}} (and likely altered by user).}

\item{columnJoinPlan}{columns to join, from \code{\link{buildJoinPlan}} (and likely altered by user).}
}
\value{
NULL if okay, else a string
}
\description{
check that a join plan is consistent with table descriptions
}
\examples{

# example data
d1 <- data.frame(id= 1:3,
                 weight= c(200, 140, 98),
                 height= c(60, 24, 12))
d2 <- data.frame(pid= 2:3,
                 weight= c(130, 110),
                 width= 1)
# get the initial description of table defs
tDesc <- rbind(tableDesription('d1', d1),
               tableDesription('d2', d2))
# declare keys (and give them consitent names)
tDesc$keys[[1]] <- list(PrimaryKey= 'id')
tDesc$keys[[2]] <- list(PrimaryKey= 'pid')
# build the column join plan
columnJoinPlan <- buildJoinPlan(tDesc)
# decide we don't want the width column
columnJoinPlan$resultColumn[columnJoinPlan$resultColumn=='width'] <- ''
# double check our plan
if(!is.null(inspectDescrAndJoinPlan(tDesc, columnJoinPlan))) {
  stop("bad join plan")
}

}
