% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinController.R
\name{key_inspector_all_cols}
\alias{key_inspector_all_cols}
\title{Return all columns as guess at preferred primary keys.}
\usage{
key_inspector_all_cols(handle)
}
\arguments{
\item{handle}{data handle}
}
\value{
map of keys to keys
}
\description{
Return all columns as guess at preferred primary keys.
}
\examples{

d <- data.frame(x=1:3, y=NA)
key_inspector_all_cols(d)

}
\seealso{
\code{tableDescription}
}
