% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinController.R
\name{key_inspector_sqlite}
\alias{key_inspector_sqlite}
\title{Return all primary key columns as guess at preferred primary keys for a SQLite handle.}
\usage{
key_inspector_sqlite(handle)
}
\arguments{
\item{handle}{data handle}
}
\value{
map of keys to keys
}
\description{
Return all primary key columns as guess at preferred primary keys for a SQLite handle.
}
\seealso{
\code{tableDescription}
}
