% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinController.R
\name{makeJoinDiagramSpec}
\alias{makeJoinDiagramSpec}
\title{Build a drawable specification of the join diagram}
\usage{
makeJoinDiagramSpec(columnJoinPlan)
}
\arguments{
\item{columnJoinPlan}{join plan}
}
\value{
mermaid diagram spec
}
\description{
Build a drawable specification of the join diagram
}
\examples{

}
