% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowscols.R
\name{moveValuesToRowsQ}
\alias{moveValuesToRowsQ}
\title{Map a set of columns to rows (query based).}
\usage{
moveValuesToRowsQ(controlTable, columnsToCopy, wideTableName, my_db, ...,
  tempNameGenerator = replyr::makeTempNameGenerator("er"),
  showQuery = FALSE)
}
\arguments{
\item{controlTable}{table specifying mapping (local data frame)}

\item{columnsToCopy}{character list of column names to copy}

\item{wideTableName}{name of table containing data to be mapped (db/Spark data)}

\item{my_db}{db handle}

\item{...}{force later arguments to be by name.}

\item{tempNameGenerator}{a tempNameGenerator from replyr::makeTempNameGenerator()}

\item{showQuery}{if TRUE print query}
}
\value{
long table built by mapping wideTable to one row per group
}
\description{
The controlTable is a table whose first column defines a group and
remaining columns define column selections for that group.   The
result of moveValuesToRowsQ() is a cross join of the controlTable
and the wideTable with what values are in the columns name by the
column selections in the controlTable given by the names in the
rows of controTable.  This is essentially a multi-column
un-pivot, gather, or moveValuesToRows.  The operation is performed
through the DBI SQL interface as a single cross join with case
statements.
}
\examples{

suppressPackageStartupMessages(library("dplyr"))
my_db <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
wideTableName <- 'dat'
d <- copy_to(my_db,
             tribble(~ID, ~v1, ~v2, ~v3, ~v4,
                       1, 101, 201, 301, 401,
                       2, 102, 202, 302, 402,
                       3, 103, 203, 303, 403),
             wideTableName, overwrite = TRUE, temporary=TRUE)
controlTable <- tribble(~group, ~col1, ~col2,
                          'aa',  'v1',  'v3',
                          'bb',  'v2',  'v4')
columnsToCopy = 'ID'
moveValuesToRowsQ(controlTable, columnsToCopy,
                  wideTableName,
                  my_db)
#  # Source:   table<er_tqbavuiflwhgu4i5v7yn_0000000001> [?? x 4]
#  # Database: sqlite 3.19.3 [:memory:]
#       ID group  col1  col2
#    <dbl> <chr> <dbl> <dbl>
#  1     1    aa   101   301
#  2     1    bb   201   401
#  3     2    aa   102   302
#  4     2    bb   202   402
#  5     3    aa   103   303
#  6     3    bb   203   403

}
\seealso{
\url{https://github.com/WinVector/cdata}, \code{\link[cdata]{moveValuesToRows}}, \code{\link[cdata]{moveValuesToColumns}}, \code{\link{moveValuesToRowsQ}}, \code{\link{moveValuesToColumnsQ}}
}
