% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stmtPlanner.R
\name{partition_mutate_nse}
\alias{partition_mutate_nse}
\title{Partition a sequence of mutate commands into longest ordered no create/use blocks.}
\usage{
partition_mutate_nse(...)
}
\arguments{
\item{...}{mutate expressions}
}
\value{

}
\description{
Partition a sequence of mutate commands into longest ordered no create/use blocks.
}
\examples{

partition_mutate_nse(a1 := 1, b1 := a1, a2 := 2, b2 := a1 + a2)

}
