% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stmtPlanner.R
\name{partition_mutate_se}
\alias{partition_mutate_se}
\title{Partition a sequence of mutate commands into longest ordered no create/use blocks.}
\usage{
partition_mutate_se(exprs)
}
\arguments{
\item{exprs}{source of mutate expressions as an assignment list}
}
\value{
ordered list of mutate_se assignment blocks
}
\description{
Partition a sequence of mutate commands into longest ordered no create/use blocks.
}
\examples{

partition_mutate_se(c("a1" := "1", "b1" := "a1", "a2" := "2", "b2" := "a1 + a2"))

}
