% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinController.R
\name{renderJoinDiagram}
\alias{renderJoinDiagram}
\title{Render a diagram spec from \code{\link{makeJoinDiagramSpec}} as a PNG graphics item.}
\usage{
renderJoinDiagram(diagramSpec, ..., pngFileName = NULL, tempDir = tempdir())
}
\arguments{
\item{diagramSpec}{diagram specification from \code{\link{makeJoinDiagramSpec}}.}

\item{...}{force later arguments to bind by name.}

\item{pngFileName}{optional, file path where to save the PNG.}

\item{tempDir}{directory to create temporary files in (not deleted by this method).}
}
\value{
png rendering of diagram
}
\description{
Requires packages \code{DiagrammeR}, \code{htmlwidgets}, \code{webshot}, and \code{magick} properly installed to use.
}
\seealso{
\code{\link{tableDescription}}, \code{\link{buildJoinPlan}}, \code{\link{makeJoinDiagramSpec}}, \code{\link{executeLeftJoinPlan}}
}
