% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underscoreReplacements.R
\name{replyr_arrange}
\alias{replyr_arrange}
\title{arrange by a single column}
\usage{
replyr_arrange(.data, colname, descending = FALSE)
}
\arguments{
\item{.data}{data object to work on}

\item{colname}{character column name}

\item{descending}{logical if true sort descending (else sort ascending)}
}
\description{
arrange by a single column
}
\examples{

d <- data.frame(Sepal_Length= c(5.8,5.7),
                Sepal_Width= c(4.0,4.4))
replyr_arrange(d, 'Sepal_Length', descending= TRUE)

}
