% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_rows.R
\name{replyr_bind_rows}
\alias{replyr_bind_rows}
\title{bind a list of items by rows}
\usage{
replyr_bind_rows(lst)
}
\arguments{
\item{lst}{list of items to combine, must be all in same dplyr data service}
}
\value{
single data item
}
\description{
bind a list of items by rows
}
\examples{

d <- data.frame(x=1:2)
replyr_bind_rows(list(d,d,d))

}

