% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colClasses.R
\name{replyr_colClasses}
\alias{replyr_colClasses}
\title{Get column classes.}
\usage{
replyr_colClasses(x)
}
\arguments{
\item{x}{tbl or item that can be coerced into such.}
}
\value{
list of column classes.
}
\description{
Get column classes.
}
\examples{

d <- data.frame(x=c(1,2))
replyr_colClasses(d)

}
