% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyToFrom.R
\name{replyr_copy_from}
\alias{replyr_copy_from}
\title{Bring remote data back as a local data frame tbl.}
\usage{
replyr_copy_from(d, maxrow = 1e+06)
}
\arguments{
\item{d}{remote dplyr data item}

\item{maxrow}{max rows to allow (stop otherwise, set to NULL to allow any size).}
}
\value{
local tbl.
}
\description{
Bring remote data back as a local data frame tbl.
}
\examples{


if (requireNamespace("RSQLite", quietly = TRUE)) {
  fnam <- tempfile(pattern = "replyr_ex2_sqlite", tmpdir = tempdir(), fileext = "sqlite3")
  my_db <- dplyr::src_sqlite(fnam, create = TRUE)
  d <- replyr_copy_to(my_db,data.frame(x=c(1,2)),'d')
  d2 <- replyr_copy_from(d)
  print(d2)
  rm(list=c('my_db','d','d2'))
  gc()
  file.remove(fnam)
}

}

