% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceName.R
\name{replyr_dataServiceName}
\alias{replyr_dataServiceName}
\title{return a cannonical name of the data service hosting a given data object}
\usage{
replyr_dataServiceName(df, expectedNames = c("src_sqlite", "spark_connection",
  "src_spark", "src_mysql", "src_postgres"))
}
\arguments{
\item{df}{data object}

\item{expectedNames}{some names to canonicalize to.}
}
\value{
cannonical service name (lenght 1 character array)
}
\description{
return a cannonical name of the data service hosting a given data object
}
\examples{

replyr_dataServiceName(data.frame(x=1))
replyr_dataServiceName(dplyr::as.tbl(data.frame(x=1)))

}

