% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.R
\name{replyr_dim}
\alias{replyr_dim}
\title{Compute dimensions of a tbl.}
\usage{
replyr_dim(x)
}
\arguments{
\item{x}{tbl or item that can be coerced into such.}
}
\value{
dimensions (including rows)
}
\description{
Compute dimensions of a tbl.
}
\examples{

d<- data.frame(x=c(1,2))
replyr_dim(d)

}

