% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupedApply.R
\name{replyr_gapply}
\alias{replyr_gapply}
\title{grouped apply}
\usage{
replyr_gapply(df, gcolumn, f, ..., ocolumn = NULL, decreasing = FALSE,
  bindrows = TRUE, maxgroups = 100)
}
\arguments{
\item{df}{remote dplyr data item}

\item{gcolumn}{grouping column}

\item{f}{transform function}

\item{...}{force later values to be bound by name}

\item{ocolumn}{ordering column (optional)}

\item{decreasing}{if TRUE sort in decreasing order by ocolumn}

\item{bindrows}{if TRUE bind the rows back into a data item, else return split list}

\item{maxgroups}{maximum number of groups to work over}
}
\value{
transformed frame
}
\description{
Partitions from by values in grouping column, applies a generic transform
to each group and then binds the groups back together.  Only advised for a
moderate number of groups and better if grouping column is an index.
This is powerfull
enough to implement "The Split-Apply-Combine Strategy for Data Analysis"
https://www.jstatsoft.org/article/view/v040i01
}
\examples{

library('dplyr')
d <- data.frame(group=c(1,1,2,2,2),
                order=c(.1,.2,.3,.4,.5),
                values=c(10,20,2,4,8))

# User supplied window functions.  They depend on known column names and
# the data back-end matching function names (as cumsum).
cumulative_sum <- . \%>\% arrange(order) \%>\% mutate(cv=cumsum(values))
sumgroup <- . \%>\% summarize(group=min(group),
                   minv=min(values),maxv=max(values))
# group=min(group) is a "pseudo aggregation" as group constant in groups.
rank_in_group <- . \%>\% mutate(constcol=1) \%>\%
          mutate(rank=cumsum(constcol)) \%>\% select(-constcol)

d \%>\% replyr_gapply('group',cumulative_sum,ocolumn='order')
d \%>\% replyr_gapply('group',sumgroup)
d \%>\% replyr_gapply('group',rank_in_group,ocolumn='order')
d \%>\% replyr_gapply('group',rank_in_group,ocolumn='order',decreasing=TRUE)

# # below only works for services which have a cumsum operator
# my_db <- dplyr::src_postgres(host = 'localhost',port = 5432,user = 'postgres',password = 'pg')
# dR <- replyr_copy_to(my_db,d,'dR')
# dR \%>\% replyr_gapply('group',cumulative_sum,ocolumn='order')
# dR \%>\% replyr_gapply('group',sumgroup)
# dR \%>\% replyr_gapply('group',rank_in_group,ocolumn='order')
# dR \%>\% replyr_gapply('group',rank_in_group,ocolumn='order',decreasing=TRUE)

}

