% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceName.R
\name{replyr_get_src}
\alias{replyr_get_src}
\title{Get the "remote data source" where a data.frame like object lives.}
\usage{
replyr_get_src(df)
}
\arguments{
\item{df}{data.frame style object}
}
\value{
source (string if data.frame, tbl, or data.table, NULL if unknown, remote source otherwise)
}
\description{
Get the "remote data source" where a data.frame like object lives.
}
\examples{

replyr_get_src(data.frame(x=1:2))

}
