% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceName.R
\name{replyr_has_table}
\alias{replyr_has_table}
\title{check for table}
\usage{
replyr_has_table(con, name)
}
\arguments{
\item{con}{connection}

\item{name}{character name to check for}
}
\value{
TRUE if table present
}
\description{
Work around connection v.s. handle issues \url{https://github.com/tidyverse/dplyr/issues/2849}
}
\examples{

if (requireNamespace("RSQLite", quietly = TRUE) &&
    requireNamespace("dbplyr", quietly = TRUE)) {
  my_db <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
  RSQLite::initExtension(my_db)
  d <- replyr_copy_to(my_db, data.frame(x=c(1,2)), 'd')
  print(d)
  print(replyr_has_table(my_db, 'd'))
  DBI::dbDisconnect(my_db)
}

}
