% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrow.R
\name{replyr_hasrows}
\alias{replyr_hasrows}
\title{Check if a table has rows.}
\usage{
replyr_hasrows(d)
}
\arguments{
\item{d}{tbl or item that can be coerced into such.}
}
\value{
number of rows
}
\description{
Check if a table has rows.
}
\examples{

d <- data.frame(x=c(1,2))
replyr_hasrows(d)

}
