% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceName.R
\name{replyr_is_local_data}
\alias{replyr_is_local_data}
\title{Test if data is local.}
\usage{
replyr_is_local_data(d)
}
\arguments{
\item{d}{data frame}
}
\value{
TRUE if local data (data.frame, tbl/tibble)
}
\description{
Test if data is local.
}
\examples{

replyr_is_local_data(data.frame(x=1:3))

}
