% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceName.R
\name{replyr_is_Spark_data}
\alias{replyr_is_Spark_data}
\title{Test if data is Spark.}
\usage{
replyr_is_Spark_data(d)
}
\arguments{
\item{d}{data frame}
}
\value{
TRUE if Spark data
}
\description{
Test if data is Spark.
}
\examples{

replyr_is_Spark_data(data.frame(x=1:3))

}
