% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.R
\name{replyr_ncol}
\alias{replyr_ncol}
\title{Compute number of columns of a data.frame (work around https://github.com/rstudio/sparklyr/issues/976 ).}
\usage{
replyr_ncol(x)
}
\arguments{
\item{x}{tbl or item that can be coerced into such.}
}
\value{
number of columns
}
\description{
Compute number of columns of a data.frame (work around https://github.com/rstudio/sparklyr/issues/976 ).
}
\examples{

d <- data.frame(x=c(1,2))
replyr_ncol(d)

}
