% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameRestrictCols.R
\name{replyr_rename}
\alias{replyr_rename}
\title{Rename a column}
\usage{
replyr_rename(.data, newName, oldName)
}
\arguments{
\item{.data}{data object to work on}

\item{newName}{character new column name}

\item{oldName}{character old column name}
}
\description{
Rename a column
}
\examples{

d <- data.frame(Sepal_Length= c(5.8,5.7),
                Sepal_Width= c(4.0,4.4),
                Species= 'setosa', rank=c(1,2))
replyr_rename(d, 'family', 'Species')

}
