% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameRestrictCols.R
\name{replyr_reverseMap}
\alias{replyr_reverseMap}
\title{Reverse a name assignment map (which are written NEWNAME=OLDNAME).}
\usage{
replyr_reverseMap(nmap)
}
\arguments{
\item{nmap}{named list mapping with keys specifying new column names, and values as original column names.}
}
\value{
inverse map
}
\description{
Reverse a name assignment map (which are written NEWNAME=OLDNAME).
}
\examples{

mp <- c(A='x', B='y')
print(mp)
replyr_reverseMap(mp)

}
\seealso{
\code{\link{let}}, \code{\link{replyr_apply_f_mapped}}, \code{\link{replyr_mapRestrictCols}}
}
