% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underscoreReplacements.R
\name{replyr_select}
\alias{replyr_select}
\title{select columns}
\usage{
replyr_select(.data, colnames)
}
\arguments{
\item{.data}{data object to work on}

\item{colnames}{character column names}
}
\description{
select columns
}
\examples{

d <- data.frame(Sepal_Length= c(5.8,5.7),
                Sepal_Width= c(4.0,4.4),
                Species= 'setosa', rank=c(1,2))
replyr_select(d, c('Sepal_Length', 'Species'))

}
