% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{replyr_str}
\alias{replyr_str}
\title{Show structure of table.}
\usage{
replyr_str(x)
}
\arguments{
\item{x}{tbl or item that can be coerced into such.}
}
\value{
summary text.
}
\description{
Show structure of table.
}
\examples{

d <- data.frame(x=c(1,2))
replyr_str(d)

}

