% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{replyr_summary}
\alias{replyr_summary}
\title{Compute usable summary of numeric columns of tbl.}
\usage{
replyr_summary(x, countUnique = TRUE)
}
\arguments{
\item{x}{tbl or item that can be coerced into such.}

\item{countUnique}{logical, if true include unique non-NA counts.}
}
\value{
summary of numeric columns.
}
\description{
Compute usable summary of numeric columns of tbl.
}
\examples{

d <- data.frame(x=c(1,2,3),y=c(3,5,NA),z=c('a',NA,'z'),
                stringsAsFactors=FALSE)
replyr_summary(d)

}

