% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colClasses.R
\name{replyr_testCols}
\alias{replyr_testCols}
\title{Run test on columns.}
\usage{
replyr_testCols(x, f)
}
\arguments{
\item{x}{tbl or item that can be coerced into such.}

\item{f}{test function (returning logical, not depending on data length).}
}
\value{
array results.
}
\description{
Run test on columns.
}
\examples{

d <- data.frame(x=c(1,2),y=c('a','b'))
replyr_testCols(d,is.numeric)

}

