% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinController.R
\name{tableDescription}
\alias{tableDescription}
\title{Build a nice description of a table.}
\usage{
tableDescription(tableName, handle, ...,
  keyInspector = key_inspector_all_cols)
}
\arguments{
\item{tableName}{name of table to add to join plan.}

\item{handle}{table or table handle to add to join plan (can already be in the plan).}

\item{...}{force later arguments to bind by name.}

\item{keyInspector}{function that determines prefered primary key set for table.}
}
\value{
table describing the data.
}
\description{
Please see \url{http://www.win-vector.com/blog/2017/05/managing-spark-data-handles-in-r/} for details.
Note: one usually needs to alter the keys column which is just populated with all columns.
}
\details{
Please see \code{vignette('DependencySorting', package = 'replyr')} and \code{vignette('joinController', package= 'replyr')} for more details.
}
\examples{

d <- data.frame(x=1:3, y=NA)
tableDescription('d', d)


}
\seealso{
\code{\link{buildJoinPlan}}, \code{\link{keysAreUnique}}, \code{\link{makeJoinDiagramSpec}}, \code{\link{executeLeftJoinPlan}}
}
