% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinController.R
\name{tableDesription}
\alias{tableDesription}
\title{Build a nice description of a table.}
\usage{
tableDesription(tableName, handle)
}
\arguments{
\item{tableName}{name of table to add to join plan.}

\item{handle}{table or table handle to add to join plan (can already be in the plan).}
}
\value{
table describing the data.
}
\description{
Please see \url{http://www.win-vector.com/blog/2017/05/managing-spark-data-handles-in-r/} for details.
Note: one usually needs to alter the keys column which is just populated with all columns.
}
\examples{

d <- data.frame(x=1:3, y=NA)
tableDesription('d', d)


}
