% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinController.R
\name{executeLeftJoinPlan}
\alias{executeLeftJoinPlan}
\title{Execute an ordered sequence of left joins.}
\usage{
executeLeftJoinPlan(tDesc, columnJoinPlan, ..., eagerCompute = TRUE,
  tempNameGenerator = makeTempNameGenerator("executeLeftJoinPlan"))
}
\arguments{
\item{tDesc}{description of tables, from \code{\link{tableDesription}} only used to map table names to data.}

\item{columnJoinPlan}{columns to join, from \code{\link{buildJoinPlan}} (and likely altered by user).  Note: no column names must intersect with names of the form \code{table_CLEANEDTABNAME_present}.}

\item{...}{force later arguments to bind by name.}

\item{eagerCompute}{logical if TRUE compute eager.}

\item{tempNameGenerator}{temp name generator produced by replyr::makeTempNameGenerator, used to record dplyr::compute() effects.}
}
\value{
joined table
}
\description{
Execute an ordered sequence of left joins.
}
\examples{


# example data
d1 <- data.frame(id= 1:3,
                 weight= c(200, 140, 98),
                 height= c(60, 24, 12))
d2 <- data.frame(pid= 2:3,
                 weight= c(130, 110),
                 width= 1)
# get the initial description of table defs
tDesc <- rbind(tableDesription('d1', d1),
               tableDesription('d2', d2))
# declare keys (and give them consitent names)
tDesc$keys[[1]] <- list(PrimaryKey= 'id')
tDesc$keys[[2]] <- list(PrimaryKey= 'pid')
# build the column join plan
columnJoinPlan <- buildJoinPlan(tDesc)
# decide we don't want the width column
columnJoinPlan$resultColumn[columnJoinPlan$resultColumn=='width'] <- ''
# double check our plan
if(!is.null(inspectDescrAndJoinPlan(tDesc, columnJoinPlan))) {
  stop("bad join plan")
}
# execute the left joins
executeLeftJoinPlan(tDesc, columnJoinPlan)

}
