% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.R
\name{\%land\%}
\alias{\%->\%}
\alias{\%->_\%}
\alias{\%land\%}
\alias{\%land_\%}
\title{Land a value to variable from a pipeline.}
\usage{
value \%land\% name

value \%->\% name

value \%->_\% name

value \%land_\% name
}
\arguments{
\item{value}{value to write}

\item{name}{variable to write to}
}
\value{
value
}
\description{
\%land\% and \%->\% copy a pipeline value to a variable on the RHS,
\%land_\% and \%->_\% copy a pipeline value to
a variable named by its RHS argument.  There is nothing these operators do
better than "->" and they are mostly just a proof of concept.
These operators use eager evaluation.
}
\details{
Technically these operators are
not "-> assignment", so they might not be specifically prohibited in an
oppugnant reading of some style guides.
}
\examples{

library("dplyr")
7 \%>\% sin() \%->\% z1
7 \%>\% sin() \%->_\% 'z2'
varname <- 'z3'
7 \%>\% sin() \%->_\% varname

}

