% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/let.R
\name{let}
\alias{let}
\title{Prepare expr for execution with name substitions specified in alias.}
\usage{
let(alias, expr)
}
\arguments{
\item{alias}{mapping from free names in expr to target names to use.}

\item{expr}{block to prepare for execution}
}
\value{
item ready to evaluate, need to apply with "()" to get the evaluation in own environemnt.
}
\description{
Code adapted from gtools::strmacro by Gregory R. Warnes (License: GPL-2, this portion also available GPL-2 to respect gtools license).
}
\examples{

library('dplyr')
d <- data.frame(Sepal_Length=c(5.8,5.7),
                Sepal_Width=c(4.0,4.4),
                Species='setosa',
                rank=c(1,2))
mapping = list(RankColumn='rank')
let(alias=mapping,
    expr={
       d \%>\% mutate(RankColumn=RankColumn-1) -> dres
    })()
print(dres)

}

