% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinController.R
\name{makeJoinDiagramSpec}
\alias{makeJoinDiagramSpec}
\title{Build a drawable specification of the join diagram}
\usage{
makeJoinDiagramSpec(columnJoinPlan, graphType = "graph LR")
}
\arguments{
\item{columnJoinPlan}{join plan}

\item{graphType}{first command on graph}
}
\value{
mermaid diagram spec
}
\description{
Build a drawable specification of the join diagram
}
\examples{

tDesc <- data.frame(tableName= c('employee',
                                 'orgtable',
                                 'revenue',
                                 'activity'),
                    handle= I(list(NULL, NULL, NULL, NULL)),
                    columns= I(list(c('id', 'date'),
                                  c('id', 'date', 'dept'),
                                  c('date', 'dept', 'rev'),
                                  c('id', 'date', 'hours'))),
                    keys =  I(list(c('id', 'date'),
                                 c('id', 'date'),
                                 c('date', 'dept'),
                                 c('id', 'date'))),
                    colClass= I(list(c('character', 'numeric'),
                                   c('character', 'character', 'character'),
                                   c('numeric', 'character', 'numeric'),
                                   c('character', 'numeric', 'numeric'))),
                    sourceClass= 'None',
                    isEmpty= FALSE,
                    stringsAsFactors = FALSE)
diagramSpec <- makeJoinDiagramSpec(buildJoinPlan(tDesc))
# DiagrammeR::mermaid(diagramSpec)

}
