% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinController.R
\name{makeJoinDiagramSpec}
\alias{makeJoinDiagramSpec}
\title{Build a drawable specification of the join diagram}
\usage{
makeJoinDiagramSpec(columnJoinPlan, ..., groupByKeys = TRUE,
  graphOpts = NULL)
}
\arguments{
\item{columnJoinPlan}{join plan}

\item{...}{force later arguments to bind by name}

\item{groupByKeys}{logical if true build key-equivilant sub-graphs}

\item{graphOpts}{options for graphViz}
}
\value{
mermaid diagram spec
}
\description{
Build a drawable specification of the join diagram
}
\examples{


# note: employeeAndDate is likely built as a cross-product
#       join of an employee table and set of dates of interest
#       before getting to the join controller step.  We call
#       such a table "row control" or "experimental design."
tDesc <- data.frame(tableName= c('employeeAndDate',
                                 'orgtable',
                                 'revenue',
                                 'activity'),
                    handle= I(list(NULL, NULL, NULL, NULL)),
                    columns= I(list(c('id', 'date'),
                                  c('id', 'date', 'dept'),
                                  c('date', 'dept', 'rev'),
                                  c('id', 'date', 'hours'))),
                    keys =  I(list(c('id', 'date'),
                                 c('id', 'date'),
                                 c('date', 'dept'),
                                 c('id', 'date'))),
                    colClass= I(list(c('character', 'numeric'),
                                   c('character', 'character', 'character'),
                                   c('numeric', 'character', 'numeric'),
                                   c('character', 'numeric', 'numeric'))),
                    sourceClass= 'None',
                    isEmpty= FALSE,
                    stringsAsFactors = FALSE)
diagramSpec <- makeJoinDiagramSpec(buildJoinPlan(tDesc))
# DiagrammeR::grViz(diagramSpec)

}
