% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempNameGenerator.R
\name{makeTempNameGenerator}
\alias{makeTempNameGenerator}
\title{Produce a temp name generator with a given prefix.}
\usage{
makeTempNameGenerator(prefix, suffix = NULL)
}
\arguments{
\item{prefix}{character, string to prefix temp names with.}

\item{suffix}{character, optional additional disambiguating breaking string.}
}
\value{
name generator function.
}
\description{
Returns a function f where f() or f(dumpList=FALSE) return
a new temporary name  f(TRUE) or f(dumpList=TRUE) returns
the list of names generated and clears the list.
}
\examples{

f <- makeTempNameGenerator('EX')
print(f())
print(f())
print(f(dumpList=TRUE))
print(f(dumpList=TRUE))

}
