% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempNameGenerator.R
\name{makeTempNameGenerator}
\alias{makeTempNameGenerator}
\title{Produce a temp name generator with a given prefix.}
\usage{
makeTempNameGenerator(prefix)
}
\arguments{
\item{prefix}{character, string to prefix temp names with.}
}
\value{
name generator function.
}
\description{
Returns a function f where: f() returns a new temporary name,
f(remove=vector) removes names in vector and returns what was removed,
f(dumpList=TRUE) returns the list of names generated and clears the list,
f(peek=TRUE) returns the list without altering anything.
}
\examples{

f <- makeTempNameGenerator('EX')
print(f())
nm2 <- f()
print(nm2)
f(remove=nm2)
print(f(dumpList=TRUE))

}
