% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowscols.R
\name{moveValuesToColumnsQ}
\alias{moveValuesToColumnsQ}
\title{Map sets rows to columns (query based).}
\usage{
moveValuesToColumnsQ(keyColumns, controlTable, tallTableName, my_db, ...,
  columnsToCopy = NULL,
  tempNameGenerator = replyr::makeTempNameGenerator("mvtcq"),
  strict = FALSE, checkNames = TRUE, showQuery = FALSE,
  defaultValue = NULL, dropDups = FALSE)
}
\arguments{
\item{keyColumns}{character list of column defining row groups}

\item{controlTable}{table specifying mapping (local data frame)}

\item{tallTableName}{name of table containing data to be mapped (db/Spark data)}

\item{my_db}{db handle}

\item{...}{force later arguments to be by name.}

\item{columnsToCopy}{character list of column names to copy}

\item{tempNameGenerator}{a tempNameGenerator from replyr::makeTempNameGenerator()}

\item{strict}{logical, if TRUE check control table contents for uniqueness}

\item{checkNames}{logical, if TRUE check names}

\item{showQuery}{if TRUE print query}

\item{defaultValue}{if not NULL literal to use for non-match values.}

\item{dropDups}{logical if TRUE supress duplicate columns (duplicate determined by name, not content).}
}
\value{
wide table built by mapping key-grouped tallTable rows to one row per group
}
\description{
Transform data facts from rows into additional columns using SQL
and controlTable.
}
\details{
This is using the theory of "fluid data"n
(\url{https://github.com/WinVector/cdata}), which includes the
principle that each data cell has coordinates independent of the
storage details and storage detail dependent coordinates (usually
row-id, column-id, and group-id) can be re-derived at will (the
other principle is that there may not be "one true preferred data
shape" and many re-shapings of data may be needed to match data to
different algorithms and methods).

The controlTable defines the names of each data element in the two notations:
the notation of the tall table (which is row oriented)
and the notation of the wide table (which is column oriented).
controlTable[ , 1] (the group label) cross colnames(controlTable)
(the column labels) are names of data cells in the long form.
controlTable[ , 2:ncol(controlTable)] (column labels)
are names of data cells in the wide form.
To get behavior similar to tidyr::gather/spread one build the control table
by running an appropiate query over the data.

Some discussion and examples can be found here:
\url{https://winvector.github.io/replyr/articles/FluidData.html} and
here \url{https://github.com/WinVector/cdata}.
}
\examples{

my_db <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")

# pivot / tidyr::spread example
# cdata version
d <- data.frame(meas= c('AUC', 'R2'), val= c(0.6, 0.2))
cdata::pivotValuesToColumns(d,
                           columnToTakeKeysFrom= 'meas',
                           columnToTakeValuesFrom= 'val',
                           rowKeyColumns= c())
# tidyr version: tidyr::spread(d, key = 'meas', value = 'val')
# replyr moveValuesToColumnsQ() version
dR <- replyr_copy_to(my_db, d, 'dR',
                     overwrite = TRUE, temporary = TRUE)
cT <- buildPivotControlTable(dR,
                             columnToTakeKeysFrom= 'meas',
                             columnToTakeValuesFrom= 'val')
replyr::moveValuesToColumnsQ(keyColumns = NULL,
                             cT, 'dR', my_db)


# non-trival transform
tallTableName <- 'dat'
d <- dplyr::copy_to(my_db,
                    dplyr::tribble(
    ~ID, ~group,       ~col1,       ~col2,
    'id1', 'aa', 'val_id1_c1', 'val_id1_c2',
    'id1', 'bb', 'val_id1_c1', 'val_id1_c3',
    'id2', 'aa', 'val_id2_c1', 'val_id2_c2',
    'id2', 'bb', 'val_id2_c1', 'val_id2_c3',
    'id3', 'aa', 'val_id3_c1', 'val_id3_c2',
    'id3', 'bb', 'val_id3_c1', 'val_id3_c3' ),
         tallTableName,
         overwrite = TRUE, temporary=TRUE)
controlTable <- dplyr::tribble(~group, ~col1, ~col2,
                                 'aa',  'c1',  'c2',
                                 'bb',  NA,    'c3')
keyColumns <- 'ID'
moveValuesToColumnsQ(keyColumns,
                     controlTable,
                     tallTableName,
                     my_db)


}
\seealso{
\code{\link[cdata]{pivotValuesToColumns}}, \code{\link{moveValuesToRowsQ}}, \code{\link{buildPivotControlTable}}
}
