% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowscols.R
\name{moveValuesToColumnsQ}
\alias{moveValuesToColumnsQ}
\title{Map sets rows to columns (query based). NOT IMPLEMENTED YET.}
\usage{
moveValuesToColumnsQ(controlTable, keyColumns, columnsToCopy, tallTableName,
  my_db, ..., tempNameGenerator = replyr::makeTempNameGenerator("er"),
  showQuery = FALSE)
}
\arguments{
\item{controlTable}{table specifying mapping (local data frame)}

\item{keyColumns}{character list of column defining row groups}

\item{columnsToCopy}{character list of column names to copy}

\item{tallTableName}{name of table containing data to be mapped (db/Spark data)}

\item{my_db}{db handle}

\item{...}{force later arguments to be by name.}

\item{tempNameGenerator}{a tempNameGenerator from replyr::makeTempNameGenerator()}

\item{showQuery}{if TRUE print query}
}
\value{
long table built by mapping tallTable to one row per group
}
\description{
The controlTable is a table whose first column defines a group and
remaining columns define column selections for that group.   The
result of moveValuesToColumnsQ() is a cross join of the controlTable
and the tallTable with what values are in the columns name by the
column selections in the controlTable given by the names in the
rows of controTable.  This is essentially a multi-column
un-pivot, gather, or moveValuesToRows.  The operation is performed
through the DBI SQL interface as a single cross join with case
statements.
}
\seealso{
\url{https://github.com/WinVector/cdata}, \code{\link[cdata]{moveValuesToRows}}, \code{\link[cdata]{moveValuesToColumns}}, \code{\link{moveValuesToRowsQ}}, \code{\link{moveValuesToColumnsQ}}
}
