% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinController.R
\name{renderJoinDiagram}
\alias{renderJoinDiagram}
\title{Render a diagram spec from \code{\link{makeJoinDiagramSpec}} as a PNG graphics item.}
\usage{
renderJoinDiagram(diagramSpec, ...)
}
\arguments{
\item{diagramSpec}{diagram specification from \code{\link{makeJoinDiagramSpec}}.}

\item{...}{force later arguments to bind by name.}
}
\value{
DiagrammeR::grViz result
}
\description{
Requires packages \code{DiagrammeR} properly installed to use.
Please see \code{vignette('DependencySorting', package = 'replyr')} and \code{vignette('joinController', package= 'replyr')} for more details.
}
\seealso{
\code{\link{tableDescription}}, \code{\link{buildJoinPlan}}, \code{\link{makeJoinDiagramSpec}}, \code{\link{executeLeftJoinPlan}}, \code{\link{convertDiagrameToPNG}}
}
