% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replyr.R
\docType{package}
\name{replyr}
\alias{replyr}
\alias{replyr-package}
\title{replyr: Diligent Use of Big Data for R}
\description{
Methods to diligently use 'dplyr' remote data sources in R ('SQL' databases,
'Spark' 2.0.0 and above). REmote PLYing of big data for R.
Adds convenience functions to make big data tasks more like
working with an in-memory R 'data.frame'.
Results do depend on which 'dplyr' data service provider used.
}
\details{
\code{replyr} helps with the following:

\itemize{
\item Summarizing remote data (via \code{replyr_summarize}).
\item Facilitating writing "source generic" code that works similarly on multiple 'dplyr' data sources.
\item Providing big data versions of functions for splitting data, binding rows, pivoting, adding row-ids, ranking, and completing experimental designs.
\item Packaging common data manipulation tasks into operators  such as the \code{\link{gapply}} function.
\item Providing support code for common \code{SparklyR} tasks, such as tracking temporary handle IDs.
}

To learn more about replyr, please start with the vignette:
\code{vignette('replyr','replyr')}
}
