% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replyr.R
\docType{package}
\name{replyr}
\alias{replyr}
\alias{replyr-package}
\title{replyr: Fluid Use of Big Data in R}
\description{
Methods to get a grip on working with remote data sources ('SQL' databases,
sparklyr' 'Spark' 2.0.0 and above) through 'dplyr'.  Adds convenience functions to make such tasks more like
working with an in-memory 'data.frame'.  Results do depend on which 'dplyr' data service you use.
}
\details{
\code{replyr} helps with the following:

\itemize{
\item Package common data manipulation tasks into operators  such as the \code{\link{gapply}} function.
\item Provide "remote data" (SQL, Spark) replacements for functions commonly used on in-memory data frames.
\item Provide bug-fixes and work-arounds for various data services.
\item Collect and document clever dplyr tricks.
}

To learn more about replyr, please start with the vignette:
\code{vignette('replyr','replyr')}
}
