% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupedApply.R
\name{replyr_arrange}
\alias{replyr_arrange}
\title{arrange df by columnName}
\usage{
replyr_arrange(df, columnName, decreasing = FALSE)
}
\arguments{
\item{df}{data}

\item{columnName}{name of column to arrange on}

\item{decreasing}{logical, if TRUE arrange decreasing}
}
\value{
df arranged
}
\description{
arrange df by columnName
}
\examples{

replyr_arrange(data.frame(x=c(1,3,2)), 'x')
replyr_arrange(data.frame(x=c(1,3,2)), 'x', decreasing=TRUE)

}
