% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_rows.R
\name{replyr_bind_rows}
\alias{replyr_bind_rows}
\title{Bind a list of items by rows (can't use dplyr::bind_rows or dplyr::combine on remote sources).  Columns are intersected.}
\usage{
replyr_bind_rows(lst, ..., useLocalMethod = TRUE, useSparkMethod = FALSE,
  eagerTempRemoval = FALSE,
  tempNameGenerator = makeTempNameGenerator("replyr_bind_rows"))
}
\arguments{
\item{lst}{list of items to combine, must be all in same dplyr data service}

\item{...}{force other arguments to be used by name}

\item{useLocalMethod}{logical if TRUE use dplyr for local data.}

\item{useSparkMethod}{logical if TRUE try to use sdf_bind_rows on Sparklyr data}

\item{eagerTempRemoval}{logical if TRUE remove temps early.}

\item{tempNameGenerator}{temp name generator produced by replyr::makeTempNameGenerator, used to record dplyr::compute() effects.}
}
\value{
single data item
}
\description{
Can't set \code{eagerTempRemoval=TRUE} on platforms that don't correctly implement \code{dplyr::compute}
(for instance \code{Sparklyr} prior to full resolution of \url{https://github.com/rstudio/sparklyr/issues/721}).
}
\examples{

d <- data.frame(x=1:2)
replyr_bind_rows(list(d,d,d,d,d), useLocalMethod= FALSE)

}
