% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_rows.R
\name{replyr_bind_rows}
\alias{replyr_bind_rows}
\title{bind a list of items by rows (can't use dplyr::bind_rows or dplyr::combine on remote sources)}
\usage{
replyr_bind_rows(lst, ...,
  tempNameGenerator = makeTempNameGenerator("replyr_bind_rows"))
}
\arguments{
\item{lst}{list of items to combine, must be all in same dplyr data service}

\item{...}{force other arguments to be used by name}

\item{tempNameGenerator}{temp name generator produced by replyr::makeTempNameGenerator, used to record dplyr::compute() effects.}
}
\value{
single data item
}
\description{
bind a list of items by rows (can't use dplyr::bind_rows or dplyr::combine on remote sources)
}
\examples{

d <- data.frame(x=1:2)
replyr_bind_rows(list(d,d,d))

}
