% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underscoreReplacements.R
\name{replyr_group_by}
\alias{replyr_group_by}
\title{group_by columns}
\usage{
replyr_group_by(.data, colnames)
}
\arguments{
\item{.data}{data.frame}

\item{colnames}{character vector of column names to group by.}
}
\value{
.data grouped by columns named in colnames
}
\description{
See also: \url{https://gist.github.com/skranz/9681509}
}
\examples{

d <- data.frame(Sepal_Length= c(5.8,5.7),
                Sepal_Width= c(4.0,4.4),
                Species= 'setosa')
replyr_group_by(d, 'Species')

}
\seealso{
\code{\link[dplyr]{group_by}}, \code{\link[dplyr]{group_by_at}}
}
