% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceName.R
\name{replyr_list_tables}
\alias{replyr_list_tables}
\title{list tables}
\usage{
replyr_list_tables(con)
}
\arguments{
\item{con}{connection}
}
\value{
list of tables names
}
\description{
Work around connection v.s. handle issues \url{https://github.com/tidyverse/dplyr/issues/2849}
}
\examples{

if (requireNamespace("RSQLite", quietly = TRUE) &&
    requireNamespace("dbplyr", quietly = TRUE)) {
   my_db <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
   RSQLite::initExtension(my_db)
  d <- replyr_copy_to(my_db, data.frame(x=c(1,2)), 'd',
                      overwrite=TRUE, temporary=TRUE)
  print(d)
  print(replyr_list_tables(my_db))
  DBI::dbDisconnect(my_db)
}

}
