% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowscols_old.R
\name{replyr_moveValuesToRows}
\alias{replyr_moveValuesToRows}
\title{Collect values found in columnsToTakeFrom as tuples (experimental, only suitable for a moderate number of columns, not fully tested on multiple data suppliers).}
\usage{
replyr_moveValuesToRows(data, nameForNewKeyColumn, nameForNewValueColumn,
  columnsToTakeFrom, ..., na.rm = FALSE, nameForNewClassColumn = NULL,
  tempNameGenerator = makeTempNameGenerator("replyr_moveValuesToRows"))
}
\arguments{
\item{data}{data.frame to work with.}

\item{nameForNewKeyColumn}{character name of column to write new keys in.}

\item{nameForNewValueColumn}{character name of column to write new values in.}

\item{columnsToTakeFrom}{character array names of columns to take values from.}

\item{...}{force later columns to bind by name.}

\item{na.rm}{logical if TRUE remove rows with NA in nameForNewValueColumn.}

\item{nameForNewClassColumn}{optional name to land original cell classes to.}

\item{tempNameGenerator}{temp name generator produced by replyr::makeTempNameGenerator, used to record dplyr::compute() effects.}
}
\value{
data item
}
\description{
Deprecated, please use \code{\link[cdata]{moveValuesToRows}} or \code{\link{moveValuesToRowsQ}}.
}
\details{
Collect values found in columnsToTakeFrom as tuples naming which column the value came from (placed in nameForNewKeyColumn)
and value found (placed in nameForNewValueColumn).  This is essentially a \code{tidyr::gather}, \code{dplyr::melt}, or anti-pivot.
Similar interface as in the \code{cdata} package (though does not perform pre/post condition checks).
}
\seealso{
\url{https://github.com/WinVector/cdata}, \code{\link[cdata]{moveValuesToRows}}, \code{\link[cdata]{moveValuesToColumns}}, \code{\link{moveValuesToRowsQ}}, \code{\link{moveValuesToColumnsQ}}, \code{\link{replyr_moveValuesToRows}}, \code{\link{replyr_moveValuesToColumns}}
}
