% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrow.R
\name{replyr_nrow}
\alias{replyr_nrow}
\title{Compute number of rows of a tbl.}
\usage{
replyr_nrow(x)
}
\arguments{
\item{x}{tbl or item that can be coerced into such.}
}
\value{
number of rows
}
\description{
Number of row in a table.  This function is not "group aware" it returns the total number of rows, not rows per dplyr group.
}
\examples{

d <- data.frame(x=c(1,2))
replyr_nrow(d)

}
