% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{replyr_quantilec}
\alias{replyr_quantilec}
\title{Compute quantiles on remote column (NA's filtered out) using cumsum.}
\usage{
replyr_quantilec(x, cname, probs = seq(0, 1, 0.25))
}
\arguments{
\item{x}{tbl or item that can be coerced into such.}

\item{cname}{column name to compute over (not 'n' or 'csum')}

\item{probs}{numeric vector of probabilities with values in [0,1].}
}
\description{
NA's filtered out and does not break ties the same as stats::quantile.
}
\examples{

d <- data.frame(xvals=rev(1:1000))
replyr_quantilec(d,'xvals')

}

